Pull develop from master on release
Instead of mergint the release into master and into
develop, this change merges the release branch into
master, taggs master, then merges master into develop.

Since we delete the release branch, the previous
behavior was causing the develop branch to have no
relatable parrents with the master branch (for a
long time). This change ensures that there is a
parrent for the develop branch back to the master
branch at every release.

This was mostly visible in a detached HEAD state
which would show a really old release as the parrent.