Make sure we always have a clean checkout
This fixes the issue where sometimes a release does not pick up new code if it was previously released and the repository was already checked out in the local cache. This used to be necessary for the old patch methodology, however it is no longer the case with the new patch branch method.